Add more content specific blocks to base.html
Block page_header:
Which is going to act as an additional header on pages that extend
this base template in case we want to put something at the top of
those pages but underneath the main header.

Block content:
A place for our main page content itself to go.

Block postloadjs:
A spot for us to put JavaScript that we want to load at the end of
the body.